code verplaatst naar mapje characters
