+++ myapp/src/routes/characters/[id]/+page.server.js
