+++ myapp/src/routes/characters/[id]/+page.svelte
